Little changes...
Maybe, I shouldn't use Menubars. It may be changed later

"spiel interface.png" is a planned vision for ui of Spiel. Of course it will be not exactly the same as on the picture. For example, I am thinking about making map not as hex grid but in some other way